mrpt::utils::CStream::Seek() now supports files larger than 2GB by using uint64_t instead of long (still see issue report for another patch required for MSVC2010) - (Closes issue #39). Thanks to Robert Schattschneider

+++ libs/base/src/utils/CFileInputStream.cpp
@@ -122,7 +122,7 @@ size_t  CFileInputStream::Write(const void *Buffer, size_t Count)
 	Method for moving to a specified position in the streamed resource.
 	 See documentation of CStream::Seek
  ---------------------------------------------------------------*/
-uint64_t CFileInputStream::Seek(long Offset, CStream::TSeekOrigin Origin)
+uint64_t CFileInputStream::Seek(uint64_t Offset, CStream::TSeekOrigin Origin)
 {
 	if (!m_if.is_open()) return 0;
 
@@ -151,7 +151,7 @@ uint64_t CFileInputStream::getTotalBytesCount()
 
 	uint64_t previousPos = getPosition();
 	uint64_t fileSize = Seek(0,sFromEnd);
-	Seek(static_cast<long>(previousPos));
+	Seek( previousPos );
 	return fileSize;
 }
