+++ libs/base/src/utils/CFileStream.cpp
@@ -142,7 +142,7 @@ size_t  CFileStream::Write(const void *Buffer, size_t Count)
 	Method for moving to a specified position in the streamed resource.
 	 See documentation of CStream::Seek
  ---------------------------------------------------------------*/
-uint64_t CFileStream::Seek(long Offset, CStream::TSeekOrigin Origin)
+uint64_t CFileStream::Seek(uint64_t Offset, CStream::TSeekOrigin Origin)
 {
 	if (!m_f.is_open()) return 0;
 
@@ -172,7 +172,7 @@ uint64_t CFileStream::getTotalBytesCount()
 
 	uint64_t previousPos = getPosition();
 	uint64_t fileSize = Seek(0,sFromEnd);
-	Seek(static_cast<long>(previousPos));
+	Seek(previousPos);
 	return fileSize;
 }
